Moved tools to utilities
