+++ src/Core/Execution/_/QueryExecuter.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using HotChocolate.Utilities;
 
 namespace HotChocolate.Execution
 {
