+++ src/Utilities/EmptyServiceProvider.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace HotChocolate.Utilities
+{
+    public sealed class EmptyServiceProvider
+        : IServiceProvider
+    {
+        public object GetService(Type serviceType) => null;
+    }
+}
