+++ src/Utilities/ServiceProviderExtensions.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace HotChocolate.Utilities
+{
+    public static class ServiceProviderExtensions
+    {
+        public static IServiceProvider Include(
+            this IServiceProvider first,
+            IServiceProvider second)
+        {
+            return new CombinedServiceProvider(first, second);
+        }
+    }
+}
